+++ scripts/macros.js
@@ -257,3 +257,4 @@ export {trickstersTransposition} from './macros/2024/classFeatures/cleric/tricke
 export {improvedDuplicity, improvedDuplicityEffect} from './macros/2024/classFeatures/cleric/trickery/improvedDuplicity.js';
 //War Domain
 export {guidedStrike} from './macros/2024/classFeatures/cleric/war/guidedStrike.js';
+export {warPriest} from './macros/2024/classFeatures/cleric/war/warPriest.js';