+++ scripts/macros/2024/classFeatures/bard/dance/inspiringMovement.js
@@ -54,7 +54,13 @@ export let inspiringMovement = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ],
     config: [