+++ scripts/macros/2024/classFeatures/bard/glamour/mantleOfInspiration.js
@@ -65,7 +65,13 @@ export let mantleOfInspiration = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ],
     config: [