+++ scripts/macros/2024/classFeatures/bard/lore/cuttingWords.js
@@ -135,7 +135,13 @@ export let cuttingWords = {
     item: [
         {
             pass: 'created',
-            macro: added
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
         }
     ],
     config: [